extract firebase ref
