+++ App/Services/index.js
@@ -1,2 +1,3 @@
 export * from './resources'
 export * from './batches'
+export * from './firebaseRef'
